requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile 
#
click-didyoumean==0.0.3
click-repl==0.1.6
click==7.0
prompt-toolkit==2.0.8     # via click-repl
python-box==3.2.3
six==1.12.0               # via click-repl, prompt-toolkit
wcwidth==0.1.7            # via prompt-toolkit
